Scope search results to the current locale and fix html lang
Every post is statically generated under all three locales, so an
unfiltered Pagefind query returned the same article once per locale —
/ja/search listed English and Spanish copies alongside the Japanese one.
Filter results to the locale of the search page.

The root layout also hardcoded <html lang="ja">, so en/es pages claimed
Japanese and Pagefind collapsed the whole site into one "ja" language
index. Make app/[locale]/layout.tsx the root layout so lang follows the
route locale; Pagefind now builds a per-language index (es, en, ja) and a
client-side script no longer has to patch documentElement.lang after
hydration.

Root metadata (metadataBase, title template, openGraph, twitter) moves
into the locale layout, where openGraph.url can be locale-specific.
