+++ web/src/app/[locale]/layout.tsx
@@ -1,6 +1,9 @@
+import "../globals.css";
+
+import { Inter, Noto_Sans_JP } from "next/font/google";
 import { notFound } from "next/navigation";
 
-import { css } from "@/styled-system/css";
+import { css, cx } from "@/styled-system/css";
 
 import { GoogleAnalytics } from "@next/third-parties/google";
 import type { Metadata } from "next";
@@ -8,8 +11,15 @@ import type { Metadata } from "next";
 import FooterBase from "@/components/FooterBase";
 import Nav from "@/components/Nav";
 import { getDictionary, isLocale, localeToOgLocale, locales } from "@/lib/i18n";
+import { SITE_URL } from "@/lib/site";
 
 import AdsScripts from "../ads-scripts";
+
+const notoSansJP = Noto_Sans_JP({
+  subsets: ["latin"],
+  weight: ["400", "500", "700", "900"],
+});
+const inter = Inter({ subsets: ["latin"], variable: "--font-inter" });
 
 const GA_TRACKING_ID =
   process.env.NODE_ENV === "production" ? "G-5X44HTLX5D" : "G-mock";
@@ -28,9 +38,21 @@ export async function generateMetadata({
 
   const dict = await getDictionary(locale);
   return {
+    metadataBase: new URL(SITE_URL),
+    title: {
+      default: "illumination-k.dev",
+      template: "%s | illumination-k.dev",
+    },
     description: dict.meta.siteDescription,
     openGraph: {
+      type: "website",
+      siteName: "illumination-k.dev",
+      url: `${SITE_URL}/${locale}`,
       locale: localeToOgLocale[locale],
+    },
+    twitter: {
+      card: "summary",
+      creator: "@illuminationK",
     },
   };
 }
@@ -47,20 +69,38 @@ export default async function LocaleLayout({
     notFound();
   }
 
+  // `lang` is emitted into the static HTML, so it also decides which Pagefind
+  // language index a page lands in. Keep it tied to the route locale.
   return (
-    <>
-      <script
-        dangerouslySetInnerHTML={{
-          __html: `document.documentElement.lang="${locale}"`,
-        }}
-      />
-      <header>
-        <Nav locale={locale} />
-      </header>
-      <main className={css({ flex: 1 })}>{children}</main>
-      <FooterBase locale={locale} />
-      <AdsScripts />
-      <GoogleAnalytics gaId={GA_TRACKING_ID} />
-    </>
+    <html lang={locale} suppressHydrationWarning>
+      <head>
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `(function(){try{var m=localStorage.getItem("color-mode");if(m==="light"||m==="dark"){document.documentElement.setAttribute("data-color-mode",m)}else if(window.matchMedia("(prefers-color-scheme:dark)").matches){document.documentElement.setAttribute("data-color-mode","dark")}else{document.documentElement.setAttribute("data-color-mode","light")}}catch(e){document.documentElement.setAttribute("data-color-mode","light")}})()`,
+          }}
+        />
+      </head>
+      <body
+        className={cx(
+          notoSansJP.className,
+          inter.variable,
+          css({
+            bg: "bg.page",
+            color: "text.primary",
+            minH: "100vh",
+            display: "flex",
+            flexDirection: "column",
+          }),
+        )}
+      >
+        <header>
+          <Nav locale={locale} />
+        </header>
+        <main className={css({ flex: 1 })}>{children}</main>
+        <FooterBase locale={locale} />
+        <AdsScripts />
+        <GoogleAnalytics gaId={GA_TRACKING_ID} />
+      </body>
+    </html>
   );
 }
